pyproject.toml
[project]
name = "frontend-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=2.1",
    "python-dotenv>=1.0.0",
    "requests>=2.28",
]
